Fix bug with dashboard initial click
When you begin on a page that is not one of the dashboard tabs
and you click on for example any status, recent will always be shown
the first time. It is not until the second click that it will switch to that status.

This is a fix for that problem.