Nodes page performance improvements
There are tow major changes in the code to improve the perofmance of the
nodes page.

First the last_report_status from the PuppetDB /nodes API is used
instead of querying the /reports endpoint and merging the reports.

Secondly queries to the event-counts endpoint will only request status
for the nodes that has been previously returned from the /nodes
endpoint. This however is limited to node counts < 1000 nodes since too
large queries will cause PuppetDB to return a query error.